requirements.txt
fastapi==0.100.0
redis==4.4.4

PyJWT==2.3.0
python-decouple==3.8